# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CalibrationReport": {
        "additionalProperties": false,
        "properties": {
          "after": {
            "$ref": "#/components/schemas/CalibrationStats"
          },
          "applied": {
            "default": false,
            "title": "Applied",
            "type": "boolean"
          },
          "backend": {
            "default": "",
            "title": "Backend",
            "type": "string"
          },
          "baseline_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Value"
          },
          "before": {
            "$ref": "#/components/schemas/CalibrationStats"
          },
          "fitted_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fitted Value"
          },
          "grid_rmse_db": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Grid Rmse Db",
            "type": "array"
          },
          "grid_values": {
            "items": {
              "type": "number"
            },
            "title": "Grid Values",
            "type": "array"
          },
          "param": {
            "enum": [
              "scattering_coefficient",
              "relative_permittivity",
              "conductivity_s_per_m"
            ],
            "title": "Param",
            "type": "string"
          },
          "per_link_after": {
            "items": {
              "$ref": "#/components/schemas/LinkError"
            },
            "title": "Per Link After",
            "type": "array"
          },
          "target_material_id": {
            "title": "Target Material Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "target_material_id",
          "param",
          "before",
          "after"
        ],
        "title": "CalibrationReport",
        "type": "object"
      },
      "CalibrationRequest": {
        "additionalProperties": false,
        "properties": {
          "apply": {
            "default": false,
            "title": "Apply",
            "type": "boolean"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Id"
          },
          "grid": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grid"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/MeasurementSample"
            },
            "minItems": 1,
            "title": "Measurements",
            "type": "array"
          },
          "param": {
            "default": "scattering_coefficient",
            "enum": [
              "scattering_coefficient",
              "relative_permittivity",
              "conductivity_s_per_m"
            ],
            "title": "Param",
            "type": "string"
          },
          "target_material_id": {
            "title": "Target Material Id",
            "type": "string"
          }
        },
        "required": [
          "measurements",
          "target_material_id"
        ],
        "title": "CalibrationRequest",
        "type": "object"
      },
...
      "DisambiguationReport": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "best_material_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Best Material Id"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/DisambiguationCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "prim_ids": {
            "items": {
              "type": "string"
            },
            "title": "Prim Ids",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "prim_ids",
          "backend"
        ],
        "title": "DisambiguationReport",
        "type": "object"
      },
      "DisambiguationRequest": {
        "additionalProperties": false,
        "description": "Which candidate RF material best explains the measurements?\n\nThe RF-sensing disambiguation step from Dai et al. (JSTEAP 2025):\nvisually identical materials (e.g. glass types spanning 2.5-23.6 dB\npenetration loss) are separated by re-simulating the measured links with\neach candidate bound to the target prims and ranking the level-aligned\nRMSE - the same error metric as parameter calibration.",
        "properties": {
          "candidate_material_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 2,
            "title": "Candidate Material Ids",
            "type": "array"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationConfig"
              },
              {
                "type": "null"
              }
            ]
          },
          "config_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Id"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/MeasurementSample"
            },
            "minItems": 1,
            "title": "Measurements",
            "type": "array"
          },
          "prim_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Prim Ids",
            "type": "array"
          }
        },
        "required": [
          "prim_ids",
          "candidate_material_ids",
          "measurements"
        ],
        "title": "DisambiguationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ImportAodtRequest": {
        "additionalProperties": false,
        "properties": {
          "kinds": {
            "items": {
              "type": "string"
            },
            "title": "Kinds",
            "type": "array"
          },
          "source_dir": {
            "title": "Source Dir",
            "type": "string"
          }
        },
        "required": [
          "source_dir"
        ],
        "title": "ImportAodtRequest",
        "type": "object"
      },
      "ImportAodtResponse": {
        "additionalProperties": false,
        "properties": {
          "imported": {
            "items": {
              "$ref": "#/components/schemas/ImportedResult"
            },
            "title": "Imported",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "title": "ImportAodtResponse",
        "type": "object"
      },
...
      "MeasurementImportRequest": {
        "additionalProperties": false,
        "properties": {
          "csv_text": {
            "title": "Csv Text",
            "type": "string"
          }
        },
        "required": [
          "csv_text"
        ],
        "title": "MeasurementImportRequest",
        "type": "object"
      },
      "MeasurementImportResponse": {
        "additionalProperties": false,
        "properties": {
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/MeasurementSample"
            },
            "title": "Measurements",
            "type": "array"
          },
          "skipped": {
            "default": 0,
            "title": "Skipped",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "title": "MeasurementImportResponse",
        "type": "object"
      },
      "MeasurementSample": {
        "additionalProperties": false,
        "properties": {
          "measured_path_gain_db": {
            "title": "Measured Path Gain Db",
            "type": "number"
          },
          "measured_rms_delay_spread_ns": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measured Rms Delay Spread Ns"
          },
          "measurement_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Measurement Id"
          },
          "rx_position": {
            "items": {
              "type": "number"
            },
            "maxItems": 3,
            "minItems": 3,
            "title": "Rx Position",
            "type": "array"
          },
          "tx_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Id"
          }
        },
        "required": [
          "rx_position",
          "measured_path_gain_db"
        ],
        "title": "MeasurementSample",
        "type": "object"
      },
...
      "SimulationConfig": {
        "additionalProperties": false,
        "description": "Full Sionna RT solver parameter surface (mirrors PathSolver/\nRadioMapSolver options so every knob is user-controllable in the UI).",
        "properties": {
          "backend": {
            "default": "auto",
            "enum": [
              "auto",
              "mock",
              "sionna"
            ],
            "title": "Backend",
            "type": "string"
          },
          "bandwidth_hz": {
            "default": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Bandwidth Hz",
            "type": "number"
          },
          "diffraction": {
            "default": false,
            "title": "Diffraction",
            "type": "boolean"
          },
          "diffraction_lit_region": {
            "default": false,
            "title": "Diffraction Lit Region",
            "type": "boolean"
          },
          "edge_diffraction": {
            "default": false,
            "title": "Edge Diffraction",
            "type": "boolean"
          },
          "engine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine"
          },
          "frequency_hz": {
            "default": 28000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Frequency Hz",
            "type": "number"
          },
          "id": {
            "default": "default",
            "title": "Id",
            "type": "string"
          },
          "los": {
            "default": true,
            "title": "Los",
            "type": "boolean"
          },
          "max_depth": {
            "default": 3,
            "maximum": 12.0,
            "minimum": 0.0,
            "title": "Max Depth",
            "type": "integer"
          },
          "name": {
            "default": "Default",
            "title": "Name",
            "type": "string"
          },
          "noise_figure_db": {
            "default": 7.0,
            "minimum": 0.0,
            "title": "Noise Figure Db",
            "type": "number"
          },
          "num_samples": {
            "default": 1000000,
            "minimum": 1.0,
            "title": "Num Samples",
            "type": "integer"
          },
          "radio_map": {
            "$ref": "#/components/schemas/RadioMapGridConfig"
          },
          "reflection": {
            "default": true,
            "title": "Reflection",
            "type": "boolean"
          },
          "refraction": {
            "default": false,
            "title": "Refraction",
            "type": "boolean"
          },
          "rx_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rx Ids"
          },
          "scattering": {
            "default": false,
            "title": "Scattering",
            "type": "boolean"
          },
          "seed": {
            "default": 42,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          },
          "synthetic_array": {
            "default": true,
            "title": "Synthetic Array",
            "type": "boolean"
          },
          "tx_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Ids"
          }
        },
        "title": "SimulationConfig",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Unified RF-visual scene authoring, RF material assignment, Sionna RT projection compilation, and simulation result APIs.",
    "title": "SionnaTwin Studio",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects/{project_id}/calibrate/disambiguate": {
      "post": {
        "description": "Rank candidate RF materials for a prim by measurement fit (the\nRF-sensing disambiguation companion to the AI suggestion flow).",
        "operationId": "disambiguate_api_projects__project_id__calibrate_disambiguate_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisambiguationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisambiguationReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Disambiguate",
        "tags": [
          "calibrate"
        ]
      }
    },
    "/api/projects/{project_id}/calibrate/materials": {
      "post": {
        "operationId": "calibrate_materials_api_projects__project_id__calibrate_materials_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calibrate Materials",
        "tags": [
          "calibrate"
        ]
      }
    },
    "/api/projects/{project_id}/calibrate/measurements": {
      "get": {
        "description": "Re-parse the stored measurement CSV (404 when none was ever imported).",
        "operationId": "get_measurements_api_projects__project_id__calibrate_measurements_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Measurements",
        "tags": [
          "calibrate"
        ]
      }
    },
    "/api/projects/{project_id}/calibrate/measurements/import-csv": {
      "post": {
        "description": "Import measurement samples from CSV text and persist the raw CSV.\n\nBad rows are skipped and counted (never fatal). The uploaded CSV is stored\nverbatim so a later GET re-parses the exact source the user provided.",
        "operationId": "import_measurements_csv_api_projects__project_id__calibrate_measurements_import_csv_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeasurementImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasurementImportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Measurements Csv",
        "tags": [
          "calibrate"
        ]
      }
    },
...
    "/api/projects/{project_id}/results/import-aodt": {
      "post": {
        "operationId": "import_aodt_api_projects__project_id__results_import_aodt_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportAodtRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportAodtResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Import Aodt",
        "tags": [
          "import"
        ]
      }
    },
...
  }
}